# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Piposmart API
  description: |
    REST API backend Piposmart (laundry management).
    Master data utama setara brief `/items` diimplementasikan sebagai `/layanan`.
    Login: `POST /auth/login`.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Health
  - name: Auth
  - name: Layanan
  - name: Outlets
  - name: Pelanggan
  - name: Karyawan
  - name: Pesanan
  - name: Dashboard

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data: {}
        meta:
          $ref: '#/components/schemas/Meta'

    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: object
          additionalProperties: true

    Meta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [owner, kasir, produksi]

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: string
          nullable: true

    LoginData:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'

    LayananRequest:
      type: object
      required: [name, kategori, harga, satuan]
      properties:
        name:
          type: string
        kategori:
          type: string
        harga:
          type: number
          exclusiveMinimum: 0
        satuan:
          type: string
        is_active:
          type: boolean

    Layanan:
      type: object
      properties:
        id:
          type: integer
        kode:
          type: string
        name:
          type: string
        kategori:
          type: string
        harga:
          type: number
        satuan:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    OutletRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: string
          nullable: true
        provinsi:
          type: string
          nullable: true
        kota:
          type: string
          nullable: true
        kecamatan:
          type: string
          nullable: true

    Outlet:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
          nullable: true
        provinsi:
          type: string
          nullable: true
        kota:
          type: string
          nullable: true
        kecamatan:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string

    PelangganRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        outlet_id:
          type: integer
          nullable: true

    Pelanggan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        outlet_id:
          type: integer
          nullable: true
        outlet_name:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string

    KaryawanCreateRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [owner, kasir, produksi]
        outlet_ids:
          type: array
          items:
            type: integer
        join_date:
          type: string
          description: YYYY-MM-DD
          nullable: true

    KaryawanUpdateRequest:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          type: string
        outlet_ids:
          type: array
          items:
            type: integer
          nullable: true
        join_date:
          type: string
          nullable: true

    Karyawan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
        outlets:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              join_date:
                type: string
                nullable: true
        created_at:
          type: string
        updated_at:
          type: string

    PesananItemInput:
      type: object
      required: [layanan_id, qty]
      properties:
        layanan_id:
          type: integer
        qty:
          type: number
          exclusiveMinimum: 0

    CreatePesananRequest:
      type: object
      required: [outlet_id, items]
      properties:
        outlet_id:
          type: integer
        pelanggan_id:
          type: integer
          nullable: true
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PesananItemInput'
        pickup_estimate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    UpdatePesananRequest:
      type: object
      properties:
        status:
          type: string
          enum: [Baru, Diproses, Siap Diambil, Selesai, Dibatalkan]
        is_lunas:
          type: boolean
        notes:
          type: string
          nullable: true
        pickup_estimate:
          type: string
          nullable: true

    Pesanan:
      type: object
      properties:
        id:
          type: integer
        kode:
          type: string
        outlet_id:
          type: integer
        outlet_name:
          type: string
        pelanggan_id:
          type: integer
          nullable: true
        pelanggan_name:
          type: string
          nullable: true
        user_id:
          type: integer
          nullable: true
        status:
          type: string
        is_lunas:
          type: boolean
        total:
          type: number
        pickup_estimate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
        layanan_summary:
          type: string
        created_at:
          type: string
        updated_at:
          type: string

    DashboardSummary:
      type: object
      properties:
        pesanan_aktif:
          type: integer
        laba_hari_ini:
          type: number
        penjualan_hari_ini:
          type: integer
        pengeluaran_hari_ini:
          type: number
        total_selesai:
          type: integer

    StatusSummary:
      type: object
      properties:
        selesai:
          type: integer
        terlambat:
          type: integer
        harus_selesai:
          type: integer

  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    QParam:
      name: q
      in: query
      schema:
        type: string
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        '200':
          description: OK
        '503':
          description: Database unavailable

  /auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
        '400':
          description: Validation error
        '409':
          description: Email conflict

  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: reynold@gmail.com
              password: password123
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/LoginData'
        '400':
          description: Validation error
        '401':
          description: Invalid credentials

  /auth/me:
    get:
      tags: [Auth]
      summary: Current user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized

  /auth/session:
    delete:
      tags: [Auth]
      summary: Logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized

  /layanan:
    get:
      tags: [Layanan]
      summary: List layanan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QParam'
        - name: kategori
          in: query
          schema:
            type: string
        - name: is_active
          in: query
          schema:
            type: string
            enum: ['true', 'false']
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: OK
    post:
      tags: [Layanan]
      summary: Create layanan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LayananRequest'
      responses:
        '201':
          description: Created
        '400':
          description: Validation error

  /layanan/{id}:
    get:
      tags: [Layanan]
      summary: Get layanan by ID
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
        '404':
          description: Not found
    put:
      tags: [Layanan]
      summary: Update layanan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LayananRequest'
      responses:
        '200':
          description: OK
        '404':
          description: Not found
    delete:
      tags: [Layanan]
      summary: Delete layanan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
        '404':
          description: Not found

  /outlets:
    get:
      tags: [Outlets]
      summary: List outlets
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: OK
    post:
      tags: [Outlets]
      summary: Create outlet
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutletRequest'
      responses:
        '201':
          description: Created

  /outlets/{id}:
    get:
      tags: [Outlets]
      summary: Get outlet
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
        '404':
          description: Not found
    put:
      tags: [Outlets]
      summary: Update outlet
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutletRequest'
      responses:
        '200':
          description: OK
    delete:
      tags: [Outlets]
      summary: Delete outlet
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK

  /pelanggan:
    get:
      tags: [Pelanggan]
      summary: List pelanggan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QParam'
        - name: outlet_id
          in: query
          schema:
            type: integer
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: OK
    post:
      tags: [Pelanggan]
      summary: Create pelanggan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PelangganRequest'
      responses:
        '201':
          description: Created

  /pelanggan/{id}:
    get:
      tags: [Pelanggan]
      summary: Get pelanggan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
    put:
      tags: [Pelanggan]
      summary: Update pelanggan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PelangganRequest'
      responses:
        '200':
          description: OK
    delete:
      tags: [Pelanggan]
      summary: Delete pelanggan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK

  /karyawan:
    get:
      tags: [Karyawan]
      summary: List karyawan (owner only)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QParam'
        - name: role
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
    post:
      tags: [Karyawan]
      summary: Create karyawan (owner only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KaryawanCreateRequest'
      responses:
        '201':
          description: Created
        '403':
          description: Forbidden

  /karyawan/{id}:
    get:
      tags: [Karyawan]
      summary: Get karyawan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
    put:
      tags: [Karyawan]
      summary: Update karyawan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KaryawanUpdateRequest'
      responses:
        '200':
          description: OK
    delete:
      tags: [Karyawan]
      summary: Delete karyawan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK

  /pesanan:
    get:
      tags: [Pesanan]
      summary: List pesanan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QParam'
        - name: status
          in: query
          schema:
            type: string
        - name: outlet_id
          in: query
          schema:
            type: integer
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: OK
    post:
      tags: [Pesanan]
      summary: Create pesanan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePesananRequest'
      responses:
        '201':
          description: Created

  /pesanan/{id}:
    get:
      tags: [Pesanan]
      summary: Get pesanan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
    put:
      tags: [Pesanan]
      summary: Update pesanan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePesananRequest'
      responses:
        '200':
          description: OK
    delete:
      tags: [Pesanan]
      summary: Delete pesanan
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK

  /transactions:
    get:
      tags: [Pesanan]
      summary: List transactions view
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QParam'
        - name: status
          in: query
          schema:
            type: string
        - name: outlet_id
          in: query
          schema:
            type: integer
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: OK

  /dashboard/summary:
    get:
      tags: [Dashboard]
      summary: Dashboard finance/order summary
      security:
        - BearerAuth: []
      parameters:
        - name: outlet_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/DashboardSummary'

  /orders/status-summary:
    get:
      tags: [Dashboard]
      summary: Order status summary
      security:
        - BearerAuth: []
      parameters:
        - name: outlet_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/StatusSummary'